fix both labs
As a result of the correct execution of
read_image_raw in section 9 of 19 variable B is overwritten. quantize_image function is not suited to work with raw data, so it returns an error in section 13 of 19. The solution is to overwrite the B variable again at the beginning of the section 13 of 19.